pyproject.toml
[tool.poetry]
name = "github-avatars-gallery-generator"
version = "0.1.0"
description = "Collect the avatars of all contributors of a Github repo and make a gallery"
authors = ["rexwangcc <rexwangcc@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
tenacity = "7.0.0"
Pillow = "7.2.0"
requests = "^2.25.1"
tqdm = "^4.60.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
isort = "^5.8.0"

[tool.poetry.scripts]
gh-gallery = 'github_avatars_gallery_generator.main:main'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
